backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
